Moved multer config to middlewares
