+++ backend/database/multerConfig.js
@@ -1,0 +1,15 @@
+import multer from "multer";
+
+// Define storage configuration
+const storage = multer.diskStorage({
+  destination: function (req, file, cb) {
+    cb(null, "uploads/");
+  },
+  filename: function (req, file, cb) {
+    cb(null, file.originalname);
+  },
+});
+
+const uploader = multer({ storage: storage });
+
+export default uploader;
